Memories-App/requirements.txt
fastapi>=0.103.0,<0.104.0
uvicorn[standard]>=0.23.0,<0.24.0
sqlalchemy>=2.0.0,<2.1.0
aiosqlite>=0.17.0
python-dotenv>=1.0.0
passlib[bcrypt]>=1.7.4
python-jose>=3.3.0
python-multipart>=0.0.5
jinja2>=3.1.2
pydantic>=2.0.0,<2.5.0
pydantic-settings>=2.0.0,<2.5.0
pytest>=7.0.0
httpx>=0.24.0
alembic>=1.11.0
